requirements.txt
marshmallow==3.7.1
PyJWT==1.7.1
python-dotenv==0.14.0
sanic==20.6.3
aiopg==1.0.0
peewee-async==0.7.0
pytest-sanic==1.6.1
pytest-mock==3.3.1